Add hack for collection published and updated
This is a hack to ensure that we have something for the atom:published
and atom:updated elements for the collection results. Each collection
should have its own published and updated timestamp and they should either
reflect the when the collection was first published on the source site
or on the data hub and when the collection was last updated (e.g, date
of the last change to any product in the collection). It's not clear
which values we should use and we don't have access to that info in the
current version, hence the hack for now.