Hack to use about field for regulating access
A similar hack is used in the nextgeoss theme. After the beta
we'll have time to work out what to do with the discoveryUserBeta
field that the SSO provides. For now, this will work.

The OAuth2 plugin sets the value of the about field to the value
of the discoveryUserBeta field that the UM returns. "true" is an
authorized user; "false" is an unauthorized beta user.